# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KeteringGo API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /restaurants:
    get:
      summary: List restaurants
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [rating_desc, name_asc] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Restaurants
        "422":
          $ref: "#/components/responses/Error"
  /restaurants/{id}:
    get:
      summary: Get restaurant details
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Restaurant details
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
    ApiErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...